Misc changes (see below for details)
Follow ups to review (by Mark F) of XD-1548 including:

- Renamed methods in ContainerListener
- Renamed configuredTimeout to timeout in ModuleDeploymentWriter

Also moved logCacheEvent from ZooKeeperConnection to ZooKeeperUtils.
