+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/zookeeper/ZooKeeperConnection.java
@@ -16,7 +16,6 @@
 
 package org.springframework.xd.dirt.zookeeper;
 
-import java.io.UnsupportedEncodingException;
 import java.util.concurrent.CopyOnWriteArraySet;
 
 import org.apache.commons.logging.Log;
@@ -24,12 +23,9 @@ import org.apache.commons.logging.LogFactory;
 import org.apache.curator.RetryPolicy;
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.CuratorFrameworkFactory;
-import org.apache.curator.framework.recipes.cache.ChildData;
-import org.apache.curator.framework.recipes.cache.PathChildrenCacheEvent;
 import org.apache.curator.framework.state.ConnectionState;
 import org.apache.curator.framework.state.ConnectionStateListener;
 import org.apache.curator.retry.ExponentialBackoffRetry;
-import org.slf4j.Logger;
 
 import org.springframework.context.SmartLifecycle;
 import org.springframework.util.Assert;
@@ -240,34 +236,6 @@ public class ZooKeeperConnection implements SmartLifecycle {
 	public void stop(Runnable callback) {
 		this.stop();
 		callback.run();
-	}
-
-	/**
-	 * Utility method to log {@link org.apache.curator.framework.recipes.cache.PathChildrenCacheEvent events}.
-	 *
-	 * @param logger logger to write to
-	 * @param event  event to log
-	 */
-	public static void logCacheEvent(Logger logger, PathChildrenCacheEvent event) {
-		ChildData data = event.getData();
-		String path = (data == null) ? "null" : data.getPath();
-		logger.info("Path cache event: {}, type: {}", path, event.getType());
-		if (data != null && logger.isTraceEnabled()) {
-			String content;
-			byte[] bytes = data.getData();
-			if (bytes == null || bytes.length == 0) {
-				content = "empty";
-			}
-			else {
-				try {
-					content = new String(data.getData(), "UTF-8");
-				}
-				catch (UnsupportedEncodingException e) {
-					content = "Could not convert content to UTF-8: " + e.toString();
-				}
-			}
-			logger.trace("Data for path {}: {}", path, content);
-		}
 	}
 
 	/**
